Pydantic 2.3 support
