+++ tests/unit_tests/test_generic_validators.py
@@ -1,366 +1,12 @@
 from uuid import uuid4
 
-import pytest
-from pydantic import ValidationError
-
-from pydantic_forms.core import FormPage, ReadOnlyField
+from pydantic_forms.core import FormPage
 from pydantic_forms.validators import (
-    Choice,
-    ContactPersonList,
     DisplaySubscription,
     Divider,
     Label,
-    ListOfOne,
-    ListOfTwo,
-    MigrationSummary,
-    OrganisationId,
-    Timestamp,
-    choice_list,
-    contact_person_list,
     migration_summary,
-    timestamp,
 )
-
-
-def test_choice():
-    class LegChoice(Choice):
-        Primary = "Primary"
-        Secondary = "Secondary"
-
-    class Form(FormPage):
-        choice: LegChoice
-
-    # Make sure label classvar is not included
-    assert len(LegChoice.__members__) == 2
-
-    # Should still count as string and enum
-    assert Form(choice="Primary").choice == "Primary"
-    assert Form(choice="Primary").choice == LegChoice.Primary
-
-    # Validation works
-    Form(choice="Primary")
-
-    with pytest.raises(ValidationError):
-        Form(choice="Wrong")
-
-
-def test_choice_default():
-    class LegChoice(Choice):
-        Primary = "Primary"
-        Secondary = "Secondary"
-
-    class Form(FormPage):
-        choice: LegChoice = LegChoice.Primary
-
-    Form(choice="Primary")
-    Form()
-    Form(choice=LegChoice.Primary)
-
-    with pytest.raises(ValidationError):
-        Form(choice="Wrong")
-
-
-def test_choice_default_str():
-    class LegChoice(Choice):
-        Primary = "Primary"
-        Secondary = "Secondary"
-
-    class Form(FormPage):
-        choice: LegChoice = "Primary"
-
-    Form(choice="Primary")
-    Form()
-    Form(choice=LegChoice.Primary)
-
-    with pytest.raises(ValidationError):
-        Form(choice="Wrong")
-
-
-def test_choice_schema():
-    class LegChoice(Choice):
-        Primary = "Primary"
-        Secondary = "Secondary"
-
-    class LegChoiceLabel(Choice):
-        Primary = ("Primary", "Primary LP")
-        Secondary = ("Secondary", "Secondary LP")
-        Tertiary = "Tertiary"
-
-    class Form(FormPage):
-        choice: LegChoice
-        choice_label: LegChoiceLabel
-
-    assert Form.schema() == {
-        "additionalProperties": False,
-        "definitions": {
-            "LegChoice": {
-                "description": "An enumeration.",
-                "enum": ["Primary", "Secondary"],
-                "title": "LegChoice",
-                "type": "string",
-            },
-            "LegChoiceLabel": {
-                "description": "An enumeration.",
-                "enum": ["Primary", "Secondary", "Tertiary"],
-                "options": {"Primary": "Primary LP", "Secondary": "Secondary LP", "Tertiary": "Tertiary"},
-                "title": "LegChoiceLabel",
-                "type": "string",
-            },
-        },
-        "properties": {
-            "choice": {"$ref": "#/definitions/LegChoice"},
-            "choice_label": {"$ref": "#/definitions/LegChoiceLabel"},
-        },
-        "required": ["choice", "choice_label"],
-        "title": "unknown",
-        "type": "object",
-    }
-
-
-def test_choice_list():
-    class LegChoice(Choice):
-        Primary = "Primary"
-        Secondary = "Secondary"
-
-    class Form(FormPage):
-        choice: choice_list(LegChoice)
-
-    # Validation works
-    Form(choice=["Primary"])
-    Form(choice=["Primary", "Primary"])
-
-    with pytest.raises(ValidationError):
-        Form(choice=["Wrong"])
-
-    with pytest.raises(ValidationError):
-        Form(choice=["Primary", "Wrong"])
-
-
-def test_choice_list_default():
-    class LegChoice(Choice):
-        Primary = "Primary"
-        Secondary = "Secondary"
-
-    class Form(FormPage):
-        choice: choice_list(LegChoice) = [LegChoice.Primary]
-
-    Form(choice=["Primary"])
-    Form()
-    Form(choice=[LegChoice.Primary])
-
-    with pytest.raises(ValidationError):
-        Form(choice=["Wrong"])
-
-
-def test_choice_list_default_str():
-    class LegChoice(Choice):
-        Primary = "Primary"
-        Secondary = "Secondary"
-
-    class Form(FormPage):
-        choice: choice_list(LegChoice) = ["Primary"]
-
-    Form(choice=["Primary"])
-    Form()
-    Form(choice=[LegChoice.Primary])
-
-    with pytest.raises(ValidationError):
-        Form(choice=["Wrong"])
-
-
-def test_choice_list_schema():
-    class LegChoice(Choice):
-        Primary = "Primary"
-        Secondary = "Secondary"
-
-    class LegChoiceLabel(Choice):
-        Primary = ("Primary", "Primary LP")
-        Secondary = ("Secondary", "Secondary LP")
-        Tertiary = "Tertiary"
-
-    class Form(FormPage):
-        choice: choice_list(LegChoice)
-        choice_label: choice_list(LegChoiceLabel)
-
-    assert Form.schema() == {
-        "additionalProperties": False,
-        "definitions": {
-            "LegChoice": {
-                "description": "An enumeration.",
-                "enum": ["Primary", "Secondary"],
-                "title": "LegChoice",
-                "type": "string",
-            },
-            "LegChoiceLabel": {
-                "description": "An enumeration.",
-                "enum": ["Primary", "Secondary", "Tertiary"],
-                "options": {"Primary": "Primary LP", "Secondary": "Secondary LP", "Tertiary": "Tertiary"},
-                "title": "LegChoiceLabel",
-                "type": "string",
-            },
-        },
-        "properties": {
-            "choice": {"items": {"$ref": "#/definitions/LegChoice"}, "type": "array"},
-            "choice_label": {
-                "items": {"$ref": "#/definitions/LegChoiceLabel"},
-                "options": {"Primary": "Primary LP", "Secondary": "Secondary LP", "Tertiary": "Tertiary"},
-                "type": "array",
-            },
-        },
-        "required": ["choice", "choice_label"],
-        "title": "unknown",
-        "type": "object",
-    }
-
-
-def test_choice_list_constraints():
-    class LegChoice(Choice):
-        Primary = "Primary"
-        Secondary = "Secondary"
-
-    class Form(FormPage):
-        choice: choice_list(LegChoice, min_items=1, unique_items=True) = ["Primary"]
-
-    m = Form(choice=[LegChoice.Primary, LegChoice.Secondary])
-    assert m.choice == [LegChoice.Primary, LegChoice.Secondary]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Form(choice=[1, 1, 1])
-    assert exc_info.value.errors() == [
-        {"loc": ("choice",), "msg": "the list has duplicated items", "type": "value_error.list.unique_items"}
-    ]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Form(choice=1)
-    assert exc_info.value.errors() == [
-        {"loc": ("choice",), "msg": "value is not a valid list", "type": "type_error.list"}
-    ]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Form(choice=[])
-    assert exc_info.value.errors() == [
-        {
-            "loc": ("choice",),
-            "msg": "ensure this value has at least 1 items",
-            "type": "value_error.list.min_items",
-            "ctx": {"limit_value": 1},
-        }
-    ]
-
-
-def test_contact_persons():
-    class Form(FormPage):
-        contact_persons: ContactPersonList
-
-    validated_data = Form(
-        contact_persons=[{"name": "test1", "email": "a@b.nl", "phone": ""}, {"name": "test2", "email": "a@b.nl"}]
-    ).dict()
-
-    expected = {
-        "contact_persons": [
-            {"email": "a@b.nl", "name": "test1", "phone": ""},
-            {"email": "a@b.nl", "name": "test2", "phone": ""},
-        ]
-    }
-    assert expected == validated_data
-
-
-def test_contact_persons_schema():
-    org_id = uuid4()
-
-    class OrgContactPersonList(ContactPersonList):
-        organisation = org_id
-        organisation_key = "key"
-        min_items = 1
-
-    class Form(FormPage):
-        contact_persons: ContactPersonList = []
-        contact_persons_org: OrgContactPersonList
-        contact_persons_org2: contact_person_list(org_id, "foo")  # noqa: F821
-
-    assert Form.schema() == {
-        "additionalProperties": False,
-        "definitions": {
-            "ContactPerson": {
-                "properties": {
-                    "email": {"format": "email", "title": "Email", "type": "string"},
-                    "name": {"format": "contactPersonName", "title": "Name", "type": "string"},
-                    "phone": {"default": "", "title": "Phone", "type": "string"},
-                },
-                "required": ["name", "email"],
-                "title": "ContactPerson",
-                "type": "object",
-            }
-        },
-        "properties": {
-            "contact_persons": {
-                "default": [],
-                "items": {"$ref": "#/definitions/ContactPerson"},
-                "organisationKey": "organisation",
-                "title": "Contact Persons",
-                "type": "array",
-            },
-            "contact_persons_org": {
-                "items": {"$ref": "#/definitions/ContactPerson"},
-                "organisationId": str(org_id),
-                "organisationKey": "key",
-                "title": "Contact Persons Org",
-                "type": "array",
-                "minItems": 1,
-            },
-            "contact_persons_org2": {
-                "items": {"$ref": "#/definitions/ContactPerson"},
-                "organisationId": str(org_id),
-                "organisationKey": "foo",
-                "title": "Contact Persons Org2",
-                "type": "array",
-            },
-        },
-        "required": ["contact_persons_org", "contact_persons_org2"],
-        "title": "unknown",
-        "type": "object",
-    }
-
-
-def test_contact_persons_nok():
-    org_id = uuid4()
-
-    class ReqContactPersonList(ContactPersonList):
-        min_items = 1
-
-    class OrgContactPersonList(ContactPersonList):
-        organisation = org_id
-        organisation_key = "key"
-
-    class Form(FormPage):
-        contact_persons: ReqContactPersonList
-        contact_persons_org: OrgContactPersonList = []
-
-    with pytest.raises(ValidationError) as error_info:
-        Form(contact_persons=[{"name": "test1", "email": "a@b"}, {"email": "a@b.nl"}])
-
-    expected = [
-        {
-            "loc": ("contact_persons", 0, "email"),
-            "msg": "value is not a valid email address",
-            "type": "value_error.email",
-        },
-        {"loc": ("contact_persons", 1, "name"), "msg": "field required", "type": "value_error.missing"},
-    ]
-    assert expected == error_info.value.errors()
-
-    with pytest.raises(ValidationError) as error_info:
-        Form(contact_persons=[])
-
-    expected = [
-        {
-            "loc": ("contact_persons",),
-            "msg": "ensure this value has at least 1 items",
-            "type": "value_error.list.min_items",
-            "ctx": {"limit_value": 1},
-        }
-    ]
-    assert expected == error_info.value.errors()
 
 
 def test_display():
@@ -417,209 +63,3 @@ def test_display_only_schema():
         "title": "unknown",
         "type": "object",
     }
-
-
-@pytest.mark.xfail(reason="fix flakey test")
-def test_read_only_field_schema():
-    class Form(FormPage):
-        read_only: int = ReadOnlyField(1, const=False)
-
-    assert Form.schema() == {
-        "title": "unknown",
-        "type": "object",
-        "properties": {
-            "read_only": {
-                "title": "Read Only",
-                "const": 1,
-                "uniforms": {"disabled": True, "value": 1},
-                "type": "integer",
-            }
-        },
-        "additionalProperties": False,
-    }
-
-
-def test_organisation_id_schema():
-    class Form(FormPage):
-        org_id: OrganisationId
-
-    assert Form.schema() == {
-        "additionalProperties": False,
-        "properties": {"org_id": {"format": "organisationId", "title": "Org Id", "type": "string"}},
-        "required": ["org_id"],
-        "title": "unknown",
-        "type": "object",
-    }
-
-
-def test_display_default():
-    some_sub_id = uuid4()
-
-    class Summary(MigrationSummary):
-        data = {"headers": ["one"]}
-
-    class Form(FormPage):
-        display_sub: DisplaySubscription = some_sub_id
-        label: Label = "bla"
-        migration_summary: Summary = "foo"
-
-    assert Form().dict() == {
-        "display_sub": some_sub_id,
-        "label": "bla",
-        "migration_summary": "foo",
-    }
-    assert Form(display_sub="").dict() == {
-        "display_sub": some_sub_id,
-        "label": "bla",
-        "migration_summary": "foo",
-    }
-
-
-def test_migration_summary_schema():
-    class Summary(MigrationSummary):
-        data = "foo"
-
-    class Form(FormPage):
-        ms1: Summary
-        ms2: migration_summary("bar")  # noqa: F821
-
-    assert Form.schema() == {
-        "additionalProperties": False,
-        "properties": {
-            "ms1": {"format": "summary", "title": "Ms1", "type": "string", "uniforms": {"data": "foo"}},
-            "ms2": {"format": "summary", "title": "Ms2", "type": "string", "uniforms": {"data": "bar"}},
-        },
-        "title": "unknown",
-        "type": "object",
-    }
-
-
-def test_timestamp_schema():
-    class Form(FormPage):
-        t1: Timestamp
-        t2: timestamp(
-            locale="nl-nl", min=1652751600, max=1672751600, date_format="DD-MM-YYYY HH:mm", time_format="HH:mm"  # noqa
-        )
-
-    assert Form.schema() == {
-        "additionalProperties": False,
-        "properties": {
-            "t1": {
-                "format": "timestamp",
-                "title": "T1",
-                "type": "number",
-                "uniforms": {
-                    "dateFormat": None,
-                    "locale": None,
-                    "max": None,
-                    "min": None,
-                    "showTimeSelect": True,
-                    "timeFormat": None,
-                },
-            },
-            "t2": {
-                "format": "timestamp",
-                "title": "T2",
-                "type": "number",
-                "uniforms": {
-                    "dateFormat": "DD-MM-YYYY HH:mm",
-                    "locale": "nl-nl",
-                    "max": 1672751600,
-                    "min": 1652751600,
-                    "showTimeSelect": True,
-                    "timeFormat": "HH:mm",
-                },
-            },
-        },
-        "required": ["t1", "t2"],
-        "title": "unknown",
-        "type": "object",
-    }
-
-
-def test_list_of_one():
-    class Form(FormPage):
-        one: ListOfOne[int]
-
-    assert Form(one=[1])
-
-    with pytest.raises(ValidationError) as error_info:
-        assert Form(one=[])
-
-    expected = [
-        {
-            "ctx": {"limit_value": 1},
-            "loc": ("one",),
-            "msg": "ensure this value has at least 1 items",
-            "type": "value_error.list.min_items",
-        }
-    ]
-    assert expected == error_info.value.errors()
-
-    with pytest.raises(ValidationError) as error_info:
-        assert Form(one=[1, 2])
-
-    expected = [
-        {
-            "ctx": {"limit_value": 1},
-            "loc": ("one",),
-            "msg": "ensure this value has at most 1 items",
-            "type": "value_error.list.max_items",
-        },
-    ]
-    assert expected == error_info.value.errors()
-
-
-def test_list_of_two():
-    class Form(FormPage):
-        two: ListOfTwo[int]
-
-    assert Form(two=[1, 2])
-
-    with pytest.raises(ValidationError) as error_info:
-        assert Form(two=[1])
-
-    expected = [
-        {
-            "ctx": {"limit_value": 2},
-            "loc": ("two",),
-            "msg": "ensure this value has at least 2 items",
-            "type": "value_error.list.min_items",
-        }
-    ]
-    assert expected == error_info.value.errors()
-
-    with pytest.raises(ValidationError) as error_info:
-        assert Form(two=[1, 2, 3])
-
-    expected = [
-        {
-            "ctx": {"limit_value": 2},
-            "loc": ("two",),
-            "msg": "ensure this value has at most 2 items",
-            "type": "value_error.list.max_items",
-        },
-    ]
-    assert expected == error_info.value.errors()
-
-
-def test_list_of_two_schema():
-    class Form(FormPage):
-        list: ListOfTwo[str]
-
-    expected = {
-        "additionalProperties": False,
-        "properties": {
-            "list": {
-                "items": {"type": "string"},
-                "title": "List",
-                "minItems": 2,
-                "maxItems": 2,
-                "type": "array",
-            }
-        },
-        "required": ["list"],
-        "title": "unknown",
-        "type": "object",
-    }
-    assert expected == Form.schema()
